Use runSpawnCommand rather than zx for media download service
