+++ packages/media-download/package.json
@@ -13,7 +13,6 @@
 	"dependencies": {
 		"@guardian/transcription-service-common": "1.0.0",
 		"@guardian/transcription-service-backend-common": "1.0.0",
-		"zx": "8.1.8",
 		"@aws-sdk/lib-storage": "3.658.1"
 	},
 	"devDependencies": {
