+++ packages/worker/src/transcribe.ts
@@ -1,4 +1,3 @@
-import { spawn } from 'child_process';
 import path from 'path';
 import { readFile } from '@guardian/transcription-service-backend-common';
 import { logger } from '@guardian/transcription-service-backend-common';
@@ -6,12 +5,7 @@ import {
 	LanguageCode,
 	languageCodes,
 } from '@guardian/transcription-service-common';
-
-interface ProcessResult {
-	code?: number;
-	stdout: string;
-	stderr: string;
-}
+import { runSpawnCommand } from '@guardian/transcription-service-backend-common/src/process';
 
 interface FfmpegResult {
 	wavPath: string;
@@ -47,47 +41,6 @@ export type WhisperBaseParams = {
 };
 
 const CONTAINER_FOLDER = '/input';
-
-const runSpawnCommand = (
-	processName: string,
-	cmd: string,
-	args: ReadonlyArray<string>,
-): Promise<ProcessResult> => {
-	return new Promise((resolve, reject) => {
-		const cp = spawn(cmd, args);
-		const stdout: string[] = [];
-		const stderr: string[] = [];
-		cp.stdout.on('data', (data) => {
-			stdout.push(data.toString());
-		});
-
-		cp.stderr.on('data', (data) => {
-			stderr.push(data.toString());
-		});
-
-		cp.on('error', (e) => {
-			stderr.push(e.toString());
-		});
-
-		cp.on('close', (code) => {
-			const result = {
-				stdout: stdout.join(''),
-				stderr: stderr.join(''),
-				code: code || undefined,
-			};
-			logger.info('Ignoring stdout to avoid logging sensitive data');
-			logger.info(`process ${processName} stderr: ${result.stderr}`);
-			if (code === 0) {
-				resolve(result);
-			} else {
-				logger.error(
-					`process ${processName} failed with code ${result.code} due to: ${result.stderr}`,
-				);
-				reject(result);
-			}
-		});
-	});
-};
 
 export const getOrCreateContainer = async (
 	tempDir: string,
